ROX-27622: Move SecurityPolicy CRD to template dir in Helm chart
To account for the move, the following additional steps were taken:

* Added a new flag to disable CRD installation: `allowUnsupportedMultipleCentralsUsingHelm=documentationWasReadAndUnderstood`
* Documented how to handle multi-tenancy with the CRD in the templates folder
* Set the flag to disable CRD installation in the operator

Note that upgrading Central on a cluster with multiple installations of
Central will likely result in a failure due to CRD management issues
described in the readme.

There was an attempt to use a Helm pre-upgrade hook to automate patching
the CRD to avoid potential upgrade failures, but unfortunately the
pre-upgrade hook is executed after Helm raises the validation failure
for the lack of CRD label and annotations.