Get rid of tag suffixes and inject `BUILD_TAG` directly
We used to rely on the Makefile which would run `git describe ...`
and append MAIN|COLLECTOR|SCANNER_TAG_SUFFIX to that to derive the
version for embedding in binaries.

Instead, we can pass the exact version via the `BUILD_TAG`
environment variable. This completely bypasses `git describe ...` and
so it's also necessary for the proper git-tag support which is
https://issues.redhat.com/browse/ROX-26026

Use of `BUILD_TAG` allows eliminating all *_TAG_SUFFIX variables
which we introduced just to make Konflux builds with `-fast` suffix
in version/tag deployable.