Fix main version embedded in the operator
While checking the operator image, I noticed that
- `operator` tag is `4.7.0-514-gba6cde5b55-fast`
- `main` tag is `4.7.x-514-gba6cde5b55-fast`
- MainVersion embedded in `operator` binary is
  `4.7.0-514-gba6cde5b55-fast`

The last thing is wrong. MainVersion must match the `main` tag, i.e.
have `.x` instead of `.0`.
It does not matter a whole lot in the grand scheme of things because
the specific images will come from the bundle CSV in case of
bundle-based deployments, but I want to fix it nevertheless.