Fix collector CPU profiler
When we removed the /module mount from the collector daemonset we
inadvertedly broke the gperftool cpu profiler, since it was using that
directory for dumping its output files. We are adding a new emptyDir
dedicated for these profiler dumps and bumping collector to the version
that uses this correct path.