doc: Clarify note about impossibility to change admin creds at runtime
The initial "while" could be interpreted as duration, equivalent to: "as
long as the password is stored in plain text, the secret is encrypted".

This is incorrect: the initial "while" is meant to highlight that even
if the plain text password file can be edited at any time, that change
won't be picked up by Central and redeployment will be needed.

This commit removes that word altogether and rewords the sentence to
prevent confusion.  # Please enter the commit message for your changes.
Lines starting