ROX-20769: Post-migration steps (#15261)
When performing an upgrade to PostgreSQL 15, some steps have to be done after the process has finished:

* Update the extensions (only pg_stat_statements now)
* Collect statistics

Do it as a migration. For those databases where upgrade was not needed (e.g. a freshly created one), extension upgrade will be a no-op with a notice:

NOTICE:  00000: version "1.12" of extension "pg_stat_statements" is already installed