ci: Remove update to collector RELEASED_VERSIONS file
This file is used to correlate collector versions to stackrox versions,
main use for it was to create the index directory for support-packages.

Since support-packages have been deprecated since 4.4, there's not much
point in keeping this file updated anymore and the only purpose it seems
to serve is taking up release engineers time ping-ponging a PR that gets
a rubber-stamped "approved" and merge from the collector team. Hence, we
are removing the update steps and removing the RELEASED_VERSIONS file
from the collector repo altogether.