Handle empty matrix in build-and-push-operator
When there's only one matrix item (RHACS_BRANDING) and when it gets
excluded (with `exclude`), the corresponding job still runs but in
no-matrix mode. It's weird but that's the way it works.

`build-and-push-operator` seems to be the only job like this and
for it the default branding STACKROX_BRANDING gets assumed which
results in building images tagged for quay.io/stackrox-io.
Subsequently, any job steps that try to work with quay.io/rhacs-eng
images fail and so I disable them with an additional clause in the
`if` condition.