Add DatabaseIsExternal to the status
Due to switching to PG15 UI would need to display a warning when Central
is still running with PG13 as an internal database (it's possible to
abuse troubleshooting mechanisms for that). The issue is that the same
warning will be displayed for ACSCS, which has different timeline for
upgrading the database.

To display the warning only on installations with an internal database,
add the corresponding field into the database status API.