ROX-28981: Add specific methods for processes and networkflows
This is previous work to #2085, it will allow collector to implement the
client side once it gets merged, then the follow up can clean up the
unused Communicate method. Finally, collector can be updated once the
clean up PR is merged to remove the method on that side as well.