fixup! ROX-14475: Upgrade to UBI9
Should resolve the permission issues we're seeing in CI. Regarding why
we do all the certificate business in the first place, I think it's
because we want to update the trusted certificates in the container with
any stackrox-generate certs+any relevant OCP certs. However, there may
be a better way to go about it. Needs further investigation and possibly
roping in other teams (e.g., Install team).