Add util to determine whether an image is built by Red Hat
Based on registry and remote heuristics.

This will be used in upcoming changes to keep the optimization at [1]
working even when a signature is present by default:

If only 1 signature integration exists, and that is the built-in red hat
signature integration, and the image is a red hat image, download the
signature. Otherwise skip.

[1] https://github.com/stackrox/stackrox/blob/e3180cf49b7e79c31640af08ea571edfc7ffdd28/pkg/images/enricher/enricher_impl.go#L712