Handle default signature integration in enricher optimization
This optimization skips signature download when no signature
integrations are present.

Adding a built-in integration would effectively disable the
optimization, and we want to avoid that.

This commit reworks the optimziation so that when only the built in
signature integration is present only signatures for Red Hat images are
downloaded - the rest will still be skipped.