Change img sig violation printer to state image is NOT verified
This could be confusing previously, because images that were not
verified by the selected integration but were successfully verified by
other signature integrations displayed messages like:

- Image signature is verified by io.stackrox.signatureintegration.a9ab4422-fa1d-4c99-a545-ea33ca57c8f8
- Image signature is verified by io.stackrox.signatureintegration.3fee323b-da48-4fe2-8041-02e0740cc4f5

There was no mention of the image not being verified by the selected
integration. Now, the message is in line with other fields, e.g. the
drop capabilities printer:

- Image signature is unverified by the specified signature integration
  (it is verified by other integration(s): <integrations that
  successfully verified the signature>).