fix: Use Negate: true in image signature query builder
Image Signature is a "required" criterion, that fires when the condition
is NOT met. In order to find violations, we search for the specified
integration ID with a negated query, so that we get what has NOT been
verified by that integration and therefore violates the criterion.

The existing value of !group.Negate likely comes from wanting to support
the opposite behavior: give the user the possibility to create a negated
signature criterion, so that it fires when the image IS verified by the
specified integration. In those cases the query would need Negate:
false.

However, the Image Signature Verified by field is configured with
negation forbidden, and therefore a group.Negate always false, see:

```go
	f.registerFieldMetadata(fieldnames.ImageSignatureVerifiedBy,
		querybuilders.ForImageSignatureVerificationStatus(),
		violationmessages.ImageContextFields,
		func(*validateConfiguration) *regexp.Regexp {
			return signatureIntegrationIDValueRegex
		},
		[]storage.EventSource{storage.EventSource_NOT_APPLICABLE},
		[]RuntimeFieldType{}, negationForbidden)  // <- see negationForbidden here
```

So !group.Negate will always be true.

On the other hand it can be hard to understand when troubleshooting
issues around the policy evaluation engine. Let's change it to true, to
make the code more readable.