Decrease markdown section level by two
Before this, each section was H3-level (starting with `###`)
that made it difficult to add any structure to the PR description, as
one needed to use H4 and H5 section headers, which were small and not
well visible.
If there is no reason for starting with H3, let's start with H1 (one
`#`) to allow devs using H2 in the description and H3 in the testing
sections.

rh-pre-commit.version: 2.3.2
rh-pre-commit.check-secrets: ENABLED