Make warning about changing integration ID uppercase
Also remove the sentence about it being referred to from other places:
it can look weird because of course it's a variable and can be referred
to from elsewhere. That referred to the default policy JSON document and
potentially UI code, but such warning does not add much value because
that is easily noticed by "git grep".