Add security cursor rules
Following best practices from Product Security [1].

The "alwaysApply" setting has been kept set to "false" rather than
"true" as suggested in the README.md [2], because according to Cursor
documentation [3] and rule-management UI that is the setup that enables
Cursor to apply rules based on the type of file being considered.

Setting "alwaysApply" to "true" would include the rules in every prompt,
even if the file type they refer to does not match what's being edited.

[1] https://gitlab.cee.redhat.com/product-security/security-cursor-rules
[2] https://gitlab.cee.redhat.com/product-security/security-cursor-rules/-/blob/cf7f6459/README.md#method-3-auto-attach-recommended
[3] https://docs.cursor.com/en/context/rules#rule-anatomy