ROX-29966: WIP: VSOCK listener
This command is intended to be deployed in a new container that's part
of the Collector DaemonSet.  It reuses Collector's certificates for
authentication.

This is a proof-of-concept to show that StackRox can utilize VSOCK to
receive data from an agent inside a virtual machine.

This code was vibe coded using Claude Code, though I did jump in
directly when fiddling with the socket and certificates.