fix: protoc version mismatch
Somehow protoc generated code with comments that indicate a slightly
different version was used than what is used in CI.  This is probably
not a big deal, but CI has validations that ensure all generated code is
properly checked in before merge.  So this is to address that test
failure.