Remove OCP version from exclusion name
Actually the latest OCP version (4.19) also uses unsigned images to hold
and extract the operator catalog, e.g.:

  registry.redhat.io/redhat/redhat-marketplace-index:v4.19

The content is then server by a container running a signed image. But
the unsigned image is run by an init container for a short time and it
would still raise violations.