Remove privileges need in compliance container
Those were needed because when calling vsock.Listen() without passing a
contextID the library figures it out automatically [1], and before doing
that it opens /dev/vsock to verify it's there [2], which is expected to be
true in both VMs and hosts.

In our case this code runs inside the compliance container of the
collector daemonset and, while we can expose /dev/vsock, opening it
requires elevated privileges.

However we don't really need the library to figure out the context ID
automatically, because we're running in the host, and there the context
ID is always 2 [3].

Therefore we can call ListenContextID directly, and skip the part that
requires /dev/vsock access.

[1] https://github.com/mdlayher/vsock/blob/908db2f631ffc247d77480009c79516a64a5003e/vsock.go#L78
[2] https://github.com/mdlayher/vsock/blob/908db2f631ffc247d77480009c79516a64a5003e/fd_linux.go#L11
[3] https://github.com/mdlayher/vsock/blob/908db2f631ffc247d77480009c79516a64a5003e/vsock.go#L29