tests: access consts via self, not class name
This way is less error-prone and allows for overriding in class hierarchy.

Note in particular that QaE2eGoCompatibilityTest had an error where it
used a constant from a different class, and its own constant was
apparently unused. I suppose this is because the author copy-pasted
another class and failed to replace all occurrences of its name. Using
self is immune to this kind of problem.