Make semaphore timeout configurable
This may be useful in scenarios where the rate of VM count to vsock
concurrency limit
(ROX_VIRTUAL_MACHINES_MAX_CONCURRENT_VSOCK_CONNECTIONS), and/or the
connections take a long time to be handled, such that the
concurrency limit is likely to be hit.

In those scenarios one may want to increase the timeout to wait a longer
time for a connection-handling slot before dropping it.

This also increases the timeout to 10 seconds, to add wiggle room for
cases where existing connections take long to get handled, due to big
index reports and/or slow connections/sensor/central.