Acquire the semaphore in the connection-handling goroutine
This way when concurrency limit is reached the main loop is not blocked,
and new connections are accepted and closed after `semaphoreTimeout`
expires. Before this, the main loop was blocked, which meant that we
would not accept new connections, leaving them in the OS accept queue,
which in turn caused that the agent would wait until its own connection
timeout expired.