Fix: new check for green field installation.
The old didn't work anymore, because it was merely checking existence of the status sub-resource. With the new setup involving the status controller the status sub-resource is populated much earlier. Let's instead test for the existence of a specific field which is only set by the helm-operator flow on successful reconciliaton.