ROX-31007: Remove initial Collector bounce (#17845)
Judging by the commentary, it's purpose was to facilitate testing when Collector loads kernel modules. For quite a while (and most certainly within the supported versions window) Collector doesn't load any modules anymore, rendering this step unnecessary. At the same time there is a frequent type of CI failures, which are caused by this step: the pattern is usually that Collector pods are getting bounced, and fail to be rescheduled.

To improve CI tests stability, remove the Collector bounce step.