Use proper in-place updating of the objects after annotation patching.
Before we were simply patching annotations on the cluster and
effectively throwing away the fresh cluster object we get after
patching. We were simply copying over the resource version to the
in-memory object we hold to allow for the later object updates
during the helm-operator reconciliation to succeed.

This doesn't seem correct anymore in our new multi-controller setup.
We have a status controller modifying the status subresource while our
helm-operator extensions execute, in particular also the annotation
patching.

With the code before this commit we were not preserving the status
modifications, instead we were pretending all is good by simply updating
the resourceVersion of our in-memory object. This caused the later
updating during the helm-operator flow to succeed, thereby throwing away
any interim status modifications.

With this change we are preserving any potential status modifications
because we let the annotation patching extension update the in-memory
object them during the patching.