ROX-28394: Skip SFA for Scanner tests on OCP 4.12 (#18141)
Similar to OCP QA tests (see #18002), skip SFA agent for 4.12 version.
Note, that it would be much better to factor out this functionality and reuse
it as needed, but since it's CI, this will take some time. In the meantime to
make the tests green, apply the simpler fix.

On top of that allow to pass --validate=false, when creating a secured cluster
CR if needed. This is required for upgrade tests, where we first install an older
version of the operator, that doesn't know anything about the new field. With
validation enabled (the default) it will fail.